move checkEndpointGetStatuscode function to main-rest package
this function should be placed inside the rest.go file to be used by other tests
